% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_boot.R
\name{b_boot}
\alias{b_boot}
\title{Perform Bayesian bootstrap}
\usage{
b_boot(data, R = 1000, fun)
}
\arguments{
\item{data}{The data to be bootstrapped}

\item{R}{Number of bootstrap replicates, default = 1000}

\item{fun}{Function to calculate the statistic of interest}
}
\value{
A vector representing the posterior of the statistic of interest (generated by fun)
}
\description{
Perform Bayesian bootstrap
}
