% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_boot.R
\name{boot_mat}
\alias{boot_mat}
\title{Generate a matrix of weights for the Bayesian Bootstrap}
\usage{
boot_mat(N, R)
}
\arguments{
\item{N}{Length of the data}

\item{R}{Number of bootstrap replicates}
}
\value{
A matrix of R rows and N columns with weights for the bootstrap. Rows should sum to 1
}
\description{
Generate a matrix of weights for the Bayesian Bootstrap
}
