% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_boot.R
\name{boots}
\alias{boots}
\alias{mu_boot}
\alias{sd_boot}
\alias{var_boot}
\title{Bayesian bootstrap helper functions}
\usage{
mu_boot(data, weights)

sd_boot(data, weights)

var_boot(data, weights)
}
\arguments{
\item{data}{Vector of data to perform the bootstrap on}

\item{weights}{vector of weights to use in the bootstrap replicate}
}
\description{
Bayesian bootstrap helper functions
}
