% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_sims.R
\name{cross_sim}
\alias{cross_sim}
\title{Generate a set of offspring from a cross.}
\usage{
cross_sim(p1, p2, r = 0.5, n_off)
}
\arguments{
\item{p1}{Matrix of parent 1 genotypes}

\item{p2}{Matrix of parent 2 genotypes}

\item{r}{Recombination rate between adjacent loci}

\item{n_off}{Number of offspring to be generated}
}
\value{
A matrix of offspring genotypes, rows are loci and columns are individuals.
}
\description{
Generate a set of offspring from a cross.
}
