% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_test.R
\name{epi_test}
\alias{epi_test}
\title{A test for epistasis.}
\source{
See Lynch and Walsh (1998) Chapter 9 for the t-test version this was adapted from.
}
\usage{
epi_test(p1, p2, f1, f2)
}
\arguments{
\item{p1}{A vector of estimated parent 1 phenotype means}

\item{p2}{A vector of estimated parent 2 phenotype means}

\item{f1}{A vector of estimated F1 phenotype means}

\item{f2}{A vector of estimated F2 phenotype means}
}
\value{
A vector of differences between the observed F2 phenotypic mean and the expected F2 phenotypic mean, a value of zero indicates an additive and/or dominance genetic architecture sufficiently explains the data. If input vectors were generated through bootstrapping the output is an estimate the sampling distribution of the difference, or an approximation of the posterior if the vectors of means were generated through Bayesian analysis.
}
\description{
A test for epistasis.
}
