% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_sims.R
\name{make_cross_parents}
\alias{make_cross_parents}
\title{Makes parents for crosses.}
\usage{
make_cross_parents(p1, p2, r = 0.5)
}
\arguments{
\item{p1}{Matrix with parent 1 genotypes, rows are loci and columns are the 2 chromosomes}

\item{p2}{Matrix with parent 2 genotypes, rows are loci and columns are the 2 chromosomes}

\item{r}{Recombination rate between adjacent loci.}
}
\value{
A matrix of genotypes for the resulting offspring. Rows are loci and there are 2 columns for chromosomes.
}
\description{
This takes in a set of parents and generates an offspring simulating recombination.
}
