% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_sims.R
\name{parent_sim}
\alias{parent_sim}
\title{Simulate Parents for Crosses}
\usage{
parent_sim(frq1, frq2, n_loc)
}
\arguments{
\item{frq1}{Vector of allele frequencies for loci in the 1st parental population}

\item{frq2}{Vector of allele frequencies for loci in the 2nd parental population}

\item{n_loc}{Number of loci being simulated}
}
\value{
A list of matrices. Each matrix contains a row for each locus and a column for each chromosome in the parent.
}
\description{
This function simulates parents from potentially diverged populations to make crosses with. It assumes that the parental populations are in Hardy-Weinberg and Linkage equilibrium and generates two parents from each of two populations.
}
