% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_sim.R
\name{pheno_sim}
\alias{pheno_sim}
\title{Simulate phenotypes from matrices of genotypes}
\usage{
pheno_sim(p1, p2, f1, f2, a, k, H2 = 0.4)
}
\arguments{
\item{p1}{Matrix of parent 1 population genotypes}

\item{p2}{Matrix of parent 2 population genotypes}

\item{f1}{Matrix of F1 population genotypes}

\item{f2}{Matrix of F2 population genotypes}

\item{a}{Vector of additive genetic effects}

\item{k}{Vector of dominance effects}

\item{H2}{Broad sense heritability}
}
\value{
A list of 4 vectors containing the P1, P2, F1, F2 phenotypes.
}
\description{
This takes in both intraparental, F1, and F2 crosses and produces phenotypes from a set of additive and dominance genetic effects. Parameters p1, p2, f1, and f2 are matrices of genotypes where each row is a locus and each column is an individual. Environmental variance is added to the resulting phenotypes by calculating the simulated phenotypic variance in the F2 and, combined with the input broad sense heritability, solving for the environmental variance.
}
