% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phenotypes}
\alias{phenotypes}
\title{Seedling Phenotypic Data}
\format{
\subsection{\code{phenotypes}}{

A data frame with 418 rows and 5 columns:
\describe{
\item{day_4}{Seedling leaf area (mm^2) 1 day after germination}
\item{day_17}{Seedling leaf area (mm^2) 14 days after germination}
\item{rgr}{Relative growth rate from day 4 to day 17, i.e. the change in log(area) between these timepoints}
\item{height}{Height (cm) measured ~4 months after germination}
\item{line}{cross type, i.e. intraparental \emph{longiflorus}, intraparental \emph{calycinus}, F1 or F2 hybrids}
}
}
}
\source{
Not yet published.
}
\usage{
phenotypes
}
\description{
Data set of phenotypes for \emph{Mimulus aurantiacus} var \emph{longiflorus} x \emph{Mimulus aurantiacus} var \emph{calycinus} crosses.
}
\keyword{datasets}
