% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lines.R
\name{plot_lines}
\alias{plot_lines}
\title{Plots estimates of phenotypic means for parental, F1, and F2 crosses}
\usage{
plot_lines(data)
}
\arguments{
\item{data}{Simulated cross data from \code{pheno_sim}}
}
\value{
A plot of the mean and 95\% CI of the mean of simulated crosses.
}
\description{
Plots estimates of phenotypic means for parental, F1, and F2 crosses
}
