% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_gen_funs.R
\name{mu_calc}
\alias{mu_calc}
\alias{var_calc}
\title{Calculate expected phenotypic variation and means caused by additive and dominant genotypic effects}
\usage{
mu_calc(frq, a, k)

var_calc(frq, a, k, mu)
}
\arguments{
\item{frq}{Vector of allele frequencies}

\item{a}{Vector of additive genetic effects (i.e. substituting an allele increases the phenotype by a value of a)}

\item{k}{Vector of dominance effects. -1 is dominance of one allele, 1 is dominance of the other, values outside would be under/over dominant, depending on fitness context}

\item{mu}{vector of phenotypic means at each locus}
}
\description{
Calculate expected phenotypic variation and means caused by additive and dominant genotypic effects
}
