% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_hist.R
\name{stacked_hist}
\alias{stacked_hist}
\title{Make Stacked Histogram}
\usage{
stacked_hist(pheno, xlab)
}
\arguments{
\item{pheno}{Phenotype to plot}

\item{xlab}{Label for the x acis}
}
\value{
A plot of stacked histogram
}
\description{
Make Stacked Histogram
}
